requirements_streamlit.txt
streamlit==1.40.0
supabase==2.10.0
openai==1.55.0
python-dotenv==1.0.1
